Increase maximum inline image size to 50kb.
Assets are supposed to be loaded in a different way than required. If we
require images, they are supposed to be handy to use, dirctly in the
code. That also means to us that we don't want to bother copying them
around, and that's especially true to libraries.

So the rule of thumb is: if the image is less than 50kb, it's OK to pass
it around handily in the code. If it's bigger, why should you REQUIRE it
anyway ? You're probably wrong !