ini2js: fix merging configs with same section
A section, e.g. [app], would overwrite the same section loaded from the
previous file and lose all of its keys.

Fixed by adding a level of iteration in the items of a section, but I
couldn't write a test for it. Karma doesn't like the shebang line.

I guess it would require moving the "ini2js" function itself into a
library but this beats writing this script as simple and inline as
possible.