Use 0.0.0.0 as the default host.
I don't see the "secutiry danger" here, but I clearly see the interest
when you need to share your interface with servers from other interfaces
or allow a distant user to display it.

Also, what's the use case to have a server on port 8000 on localhost and
another one on, say, wlan0 interface ?

If someone needs that, they's override the settings, but I feel this is
a better default.